+++ test/replication-cache/check.test.js
@@ -42,8 +42,8 @@ describe("Check", () => {
       await wait(2 * interval);
       expect(await cds.replicationCache.size()).toBe(24576);
 
-      const tenant = await cds.replicationCache.cache.get(undefined);
-      expect(tenant.cache.get("test.Books").status).toBe("READY");
+      const tenant = await cds.replicationCache.entries.get(undefined);
+      expect(tenant.entries.get("test.Books").status).toBe("READY");
 
       result = await tx.run(SELECT.from("test.Authors", ["ID", "name"]));
       expect(result.length).toBe(100);
@@ -59,8 +59,8 @@ describe("Check", () => {
       await wait(2 * interval);
       expect(await cds.replicationCache.size()).toBe(20480);
 
-      expect(tenant.cache.get("test.Books").status).toBe("OPEN");
-      expect(tenant.cache.get("test.Authors").status).toBe("READY");
+      expect(tenant.entries.get("test.Books").status).toBe("OPEN");
+      expect(tenant.entries.get("test.Authors").status).toBe("READY");
     });
   });
 
@@ -84,9 +84,9 @@ describe("Check", () => {
     await wait(2 * interval);
     expect(await cds.replicationCache.size()).toBe(45056);
 
-    const tenant = await cds.replicationCache.cache.get(undefined);
-    expect(tenant.cache.get("test.Books").status).toBe("READY");
-    expect(tenant.cache.get("test.Books.texts").status).toBe("READY");
+    const tenant = await cds.replicationCache.entries.get(undefined);
+    expect(tenant.entries.get("test.Books").status).toBe("READY");
+    expect(tenant.entries.get("test.Books.texts").status).toBe("READY");
 
     response = await GET("/odata/v4/test/Authors", {
       headers: {
@@ -107,7 +107,7 @@ describe("Check", () => {
     await wait(2 * interval);
     expect(await cds.replicationCache.size()).toBe(40960);
 
-    expect(tenant.cache.get("test.Books").status).toBe("OPEN");
-    expect(tenant.cache.get("test.Authors").status).toBe("READY");
+    expect(tenant.entries.get("test.Books").status).toBe("OPEN");
+    expect(tenant.entries.get("test.Authors").status).toBe("READY");
   });
 });